Refactored the code by dividing the main script into 3 modules, added .gitignore
