+++ gui/wvs_gui.py
@@ -19,7 +19,7 @@ class ScannerThread(QThread):
 
     def run(self):
         try:
-            cmd = ["python", "web_scanner.py", self.url]
+            cmd = ["python", "scanner.py", self.url]
             if self.config_path:
                 cmd.extend(["-c", self.config_path])
