+++ scanner.py
@@ -1,0 +1,81 @@
+import argparse
+from playwright.sync_api import sync_playwright
+from vulnerabilities import (check_http_headers, find_login_page, 
+                             test_default_credentials, find_input_elements, extract_links)
+from utils import load_config
+
+class WebVulnerabilityScanner:
+    def __init__(self, target_url, config=None):
+        self.target_url = target_url
+        self.config = config or {}
+
+        self.tests = self.config.get("tests", {
+            "check_http_headers": True,
+            "test_default_credentials": True,
+            "find_input_elements": True,
+            "test_xss_payloads": True,
+            "extract_links": True
+        })
+        self.xss_payloads = self.config.get("xss_payloads", [
+            "<script>alert('XSS')</script>",
+            "<img src=x onerror=alert('XSS')>",
+            "<svg/onload=alert('XSS')>"
+        ])
+        self.default_credentials = self.config.get("default_credentials", [
+            ("admin", "admin"),
+            ("admin", "password"),
+            ("admin", "1234"),
+            ("user", "user"),
+            ("root", "toor"),
+        ])
+
+    def start_scan(self):
+        """Run vulnerability scans."""
+        print(f"[*] Starting scan for {self.target_url}...")
+        with sync_playwright() as p:
+            browser = p.chromium.launch(headless=True)
+            context = browser.new_context()
+            page = context.new_page()
+
+            try:
+                page.goto(self.target_url)
+
+                if self.tests.get("check_http_headers"):
+                    check_http_headers(page, self.target_url)
+
+                login_form = None
+                if self.tests.get("test_default_credentials"):
+                    login_form = find_login_page(page)
+                    if login_form:
+                        test_default_credentials(page, login_form, self.default_credentials, self.target_url)
+
+                links = []
+                if self.tests.get("extract_links"):
+                    links = extract_links(page, self.target_url)
+
+                if self.tests.get("find_input_elements"):
+                    for link in links:
+                        print(f"\n[*] Scanning linked route: {link}")
+                        page.goto(link)
+                        if self.tests.get("test_xss_payloads"):
+                            find_input_elements(page, self.xss_payloads)
+                        else:
+                            find_input_elements(page)
+
+            except Exception as e:
+                print(f"[!] Error during scan: {e}")
+            finally:
+                browser.close()
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description="Web Vulnerability Scanner")
+    parser.add_argument("url", help="Target URL to scan (e.g., http://example.com)")
+    parser.add_argument("-c", "--config", help="Path to the configuration file (JSON format)", required=False)
+    args = parser.parse_args()
+
+    config = {}
+    if args.config:
+        config = load_config(args.config)
+
+    scanner = WebVulnerabilityScanner(args.url, config=config)
+    scanner.start_scan()
