+++ utils.py
@@ -1,0 +1,10 @@
+import json
+
+def load_config(config_path):
+    """Load configuration from a JSON file."""
+    try:
+        with open(config_path, 'r') as f:
+            return json.load(f)
+    except Exception as e:
+        print(f"[!] Error loading configuration file: {e}")
+        return {}
