+++ vulnerabilities.py
@@ -1,0 +1,192 @@
+from urllib.parse import urljoin, urlparse, parse_qs
+
+def check_http_headers(page, target_url):
+    """Check for security headers."""
+    print("[*] Checking HTTP headers...")
+    try:
+        response = page.context.request.get(target_url)
+        headers = response.headers
+        missing_headers = []
+
+        # Check common security headers
+        if 'x-content-type-options' not in headers:
+            missing_headers.append('X-Content-Type-Options')
+        if 'x-frame-options' not in headers:
+            missing_headers.append('X-Frame-Options')
+        if 'content-security-policy' not in headers:
+            missing_headers.append('Content-Security-Policy')
+
+        if missing_headers:
+            print(f"[-] Missing security headers: {', '.join(missing_headers)}")
+        else:
+            print("[+] All common security headers are present!")
+    except Exception as e:
+        print(f"[!] Error checking headers: {e}")
+
+def find_login_page(page):
+    """Detect if the page is a login page."""
+    print("[*] Checking for login page...")
+    try:
+        forms = page.locator("form")
+        for i in range(forms.count()):
+            form = forms.nth(i)
+            if form.locator("input[type='password']").count() > 0:
+                print("[+] Login form detected!")
+                return form
+        print("[-] No login form found.")
+        return None
+    except Exception as e:
+        print(f"[!] Error detecting login page: {e}")
+        return None
+
+def test_default_credentials(page, login_form, default_credentials, target_url):
+    """Try default credentials on the detected login form."""
+    print("[*] Testing default credentials...")
+    for username, password in default_credentials:
+        try:
+            print(f"  [*] Trying username: '{username}', password: '{password}'")
+
+            # Fill in credentials
+            login_form.locator("input[type='text'], input[name='username'], input[name='user']").fill(username)
+            login_form.locator("input[type='password']").fill(password)
+
+            # Submit the form
+            login_form.locator("button[type='submit'], input[type='submit']").click()
+
+            # Check if login was successful by monitoring redirection
+            if page.url != target_url:
+                print(f"[+] Login successful with username: '{username}', password: '{password}'")
+                return True
+        except Exception as e:
+            print(f"  [-] Error testing credentials: {e}")
+
+    print("[-] Default credentials did not work.")
+    return False
+
+def find_input_elements(page, payloads=None):
+    """Extract and test all input elements."""
+    print("[*] Searching for input elements on the target page...")
+    try:
+        # Locate input-like elements
+        input_locator = page.locator("input, textarea, form, select")
+        input_count = input_locator.count()
+
+        print(f"[+] Found {input_count} input-like element(s)!")
+        for i in range(input_count):
+            element = input_locator.nth(i)
+
+            # Check element visibility and editability
+            if not element.is_visible():
+                print(f"  [-] Skipping hidden element at index {i}")
+                continue
+            if not element.is_editable():
+                print(f"  [-] Skipping non-editable element at index {i}")
+                continue
+
+            # Get element attributes for identification
+            try:
+                tag_name = element.evaluate("el => el.tagName.toLowerCase()")
+                element_type = element.get_attribute("type") or "text"
+
+                if tag_name == "input" and element_type in  ["checkbox", "radio", "submit", "button"]:
+                    print(f"  [-] Skipping unsupported input type: {element_type}")
+                    continue
+
+                element_name = element.get_attribute("name") or "Unnamed"
+                placeholder = element.get_attribute("placeholder") or "None"
+                value = element.get_attribute("value") or "None"
+
+            except Exception as attr_error:
+                print(f"  [-] Failed to extract element attributes: {attr_error}")
+                continue
+
+            print(f"  [{i + 1}] Tag: {tag_name}, Type: {element_type}, Name: {element_name}, "
+                f"Placeholder: {placeholder}, Value: {value}")
+
+            if tag_name in ["input", "textarea"] and payloads:
+                test_xss_payloads(page, element, payloads)
+
+    except Exception as e:
+        print(f"[!] Error finding input elements: {e}")
+
+def test_xss_payloads(page, input_element, payloads):
+    """Test input elements for XSS vulnerabilities."""
+    print("[*] Testing XSS payloads...")
+    try:
+        for payload in payloads:
+            print(f"  [*] Injecting payload: {payload}")
+            try:
+                input_element.fill(payload)
+            except Exception as fill_error:
+                print(f"  [-] Could not fill input with payload: {fill_error}")
+                continue
+
+            # Check for reflected payload in the page content
+            try:
+                page_content = page.content()
+                if payload in page_content:
+                    print(f"[!] Reflected XSS detected with payload: {payload}")
+                    return True
+            except Exception as content_check_error:
+                print(f"  [-] Page content check failed: {content_check_error}")
+
+            # Test if the payload triggered an alert
+            try:
+                alert_triggered = page.evaluate(
+                    f"""
+                    () => {{
+                        try {{
+                            let triggered = false;
+                            window.alert = () => {{ triggered = true; }};
+                            document.querySelector("input").value = "{payload}";
+                            return triggered;
+                        }} catch (e) {{
+                            return false;
+                        }}
+                    }}
+                    """
+                )
+                if alert_triggered:
+                    print(f"[!] XSS vulnerability detected with payload: {payload}")
+                    return True
+            except Exception as alert_check_error:
+                print(f"  [-] Alert check failed: {alert_check_error}")
+
+        print("[+] No XSS vulnerabilities detected.")
+        return False
+
+    except Exception as e:
+        print(f"[!] Error during XSS testing: {e}")
+
+def extract_links(page, target_url):
+    """Extract all links from the current page, avoiding external redirects."""
+    print("[*] Extracting links from the page...")
+    try:
+        link_locator = page.locator("a")
+        link_count = link_locator.count()
+        links = set()
+
+        for i in range(link_count):
+            href = link_locator.nth(i).get_attribute("href")
+            if href:
+                # Normalize and join relative URLs with the base URL
+                full_url = urljoin(target_url, href)
+                parsed_url = urlparse(full_url)
+                query_params = parse_qs(parsed_url.query)
+
+                # Check if it's an internal link or a redirect to an external domain
+                if parsed_url.netloc == urlparse(target_url).netloc:
+                    # Avoid links with redirect query params pointing to external domains
+                    if "to" in query_params:
+                        redirect_target = query_params["to"][0]
+                        if urlparse(redirect_target).netloc != urlparse(target_url).netloc:
+                            print(f"[!] Skipping external redirect: {full_url}")
+                            continue
+
+                    links.add(full_url)
+
+        print(f"[+] Found {len(links)} unique internal links.")
+        return links
+    except Exception as e:
+        print(f"[!] Error extracting links: {e}")
+        return set()
