--- web_scanner.py
@@ -1,277 +1,0 @@
-import argparse
-from urllib.parse import urljoin, urlparse, parse_qs
-from playwright.sync_api import sync_playwright
-import json
-
-class WebVulnerabilityScanner:
-    def __init__(self, target_url, config=None):
-        self.target_url = target_url
-        self.config = config or {}
-
-        self.tests = self.config.get("tests", {
-            "check_http_headers": True,
-            "test_default_credentials": True,
-            "find_input_elements": True,
-            "test_xss_payloads": True,
-            "extract_links": True
-        })
-        self.xss_payloads = self.config.get("xss_payloads", [
-            "<script>alert('XSS')</script>",
-            "<img src=x onerror=alert('XSS')>",
-            "<svg/onload=alert('XSS')>"
-        ])
-        self.default_credentials = self.config.get("default_credentials", [
-            ("admin", "admin"),
-            ("admin", "password"),
-            ("admin", "1234"),
-            ("user", "user"),
-            ("root", "toor"),
-        ])
-
-    def check_http_headers(self, page):
-        """Check for security headers."""
-        print("[*] Checking HTTP headers...")
-        try:
-            response = page.context.request.get(self.target_url)
-            headers = response.headers
-            missing_headers = []
-
-            # Check common security headers
-            if 'x-content-type-options' not in headers:
-                missing_headers.append('X-Content-Type-Options')
-            if 'x-frame-options' not in headers:
-                missing_headers.append('X-Frame-Options')
-            if 'content-security-policy' not in headers:
-                missing_headers.append('Content-Security-Policy')
-
-            if missing_headers:
-                print(f"[-] Missing security headers: {', '.join(missing_headers)}")
-            else:
-                print("[+] All common security headers are present!")
-        except Exception as e:
-            print(f"[!] Error checking headers: {e}")
-
-    def find_login_page(self, page):
-        """Detect if the page is a login page."""
-        print("[*] Checking for login page...")
-        try:
-            forms = page.locator("form")
-            for i in range(forms.count()):
-                form = forms.nth(i)
-                if form.locator("input[type='password']").count() > 0:
-                    print("[+] Login form detected!")
-                    return form
-            print("[-] No login form found.")
-            return None
-        except Exception as e:
-            print(f"[!] Error detecting login page: {e}")
-            return None
-
-    def test_default_credentials(self, page, login_form):
-        """Try default credentials on the detected login form."""
-        print("[*] Testing default credentials...")
-        for username, password in self.default_credentials:
-            try:
-                print(f"  [*] Trying username: '{username}', password: '{password}'")
-
-                # Fill in credentials
-                login_form.locator("input[type='text'], input[name='username'], input[name='user']").fill(username)
-                login_form.locator("input[type='password']").fill(password)
-
-                # Submit the form
-                login_form.locator("button[type='submit'], input[type='submit']").click()
-
-                # Check if login was successful by monitoring redirection
-                if page.url != self.target_url:
-                    print(f"[+] Login successful with username: '{username}', password: '{password}'")
-                    return True
-            except Exception as e:
-                print(f"  [-] Error testing credentials: {e}")
-
-        print("[-] Default credentials did not work.")
-        return False
-
-    def find_input_elements(self, page):
-        """Extract and test all input elements."""
-        print("[*] Searching for input elements on the target page...")
-        try:
-            # Locate input-like elements
-            input_locator = page.locator("input, textarea, form, select")
-            input_count = input_locator.count()
-
-            print(f"[+] Found {input_count} input-like element(s)!")
-            for i in range(input_count):
-                element = input_locator.nth(i)
-
-                # Check element visibility and editability
-                if not element.is_visible():
-                    print(f"  [-] Skipping hidden element at index {i}")
-                    continue
-                if not element.is_editable():
-                    print(f"  [-] Skipping non-editable element at index {i}")
-                    continue
-
-                # Get element attributes for identification
-                try:
-                    tag_name = element.evaluate("el => el.tagName.toLowerCase()")
-                    element_type = element.get_attribute("type") or "text"
-
-                    if tag_name == "input" and element_type in  ["checkbox", "radio", "submit", "button"]:
-                        print(f"  [-] Skipping unsupported input type: {element_type}")
-                        continue
-
-                    element_name = element.get_attribute("name") or "Unnamed"
-                    placeholder = element.get_attribute("placeholder") or "None"
-                    value = element.get_attribute("value") or "None"
-
-                except Exception as attr_error:
-                    print(f"  [-] Failed to extract element attributes: {attr_error}")
-                    continue
-
-                print(f"  [{i + 1}] Tag: {tag_name}, Type: {element_type}, Name: {element_name}, "
-                    f"Placeholder: {placeholder}, Value: {value}")
-                
-                if tag_name in ["input", "textarea"]:
-                    self.test_xss_payloads(page, element)
-
-        except Exception as e:
-            print(f"[!] Error finding input elements: {e}")
-
-    def test_xss_payloads(self, page, input_element):
-        """Test input elements for XSS vulnerabilities."""
-        print("[*] Testing XSS payloads...")
-        try:
-            for payload in self.xss_payloads:
-                print(f"  [*] Injecting payload: {payload}")
-                try:
-                    input_element.fill(payload)
-                except Exception as fill_error:
-                    print(f"  [-] Could not fill input with payload: {fill_error}")
-                    continue
-
-                # Check for reflected payload in the page content
-                try:
-                    page_content = page.content()
-                    if payload in page_content:
-                        print(f"[!] Reflected XSS detected with payload: {payload}")
-                        return True
-                except Exception as content_check_error:
-                    print(f"  [-] Page content check failed: {content_check_error}")
-
-                # Test if the payload triggered an alert
-                try:
-                    alert_triggered = page.evaluate(
-                        f"""
-                        () => {{
-                            try {{
-                                let triggered = false;
-                                window.alert = () => {{ triggered = true; }};
-                                document.querySelector("input").value = "{payload}";
-                                return triggered;
-                            }} catch (e) {{
-                                return false;
-                            }}
-                        }}
-                        """
-                    )
-                    if alert_triggered:
-                        print(f"[!] XSS vulnerability detected with payload: {payload}")
-                        return True
-                except Exception as alert_check_error:
-                    print(f"  [-] Alert check failed: {alert_check_error}")                
-
-            print("[+] No XSS vulnerabilities detected.")
-            return False
-
-        except Exception as e:
-            print(f"[!] Error during XSS testing: {e}")
-
-    def extract_links(self, page):
-        """Extract all links from the current page, avoiding external redirects."""
-        print("[*] Extracting links from the page...")
-        try:
-            link_locator = page.locator("a")
-            link_count = link_locator.count()
-            links = set()
-
-            for i in range(link_count):
-                href = link_locator.nth(i).get_attribute("href")
-                if href:
-                    # Normalize and join relative URLs with the base URL
-                    full_url = urljoin(self.target_url, href)
-                    parsed_url = urlparse(full_url)
-                    query_params = parse_qs(parsed_url.query)
-
-                    # Check if it's an internal link or a redirect to an external domain
-                    if parsed_url.netloc == urlparse(self.target_url).netloc:
-                        # Avoid links with redirect query params pointing to external domains
-                        if "to" in query_params:
-                            redirect_target = query_params["to"][0]
-                            if urlparse(redirect_target).netloc != urlparse(self.target_url).netloc:
-                                print(f"[!] Skipping external redirect: {full_url}")
-                                continue
-
-                        links.add(full_url)
-
-            print(f"[+] Found {len(links)} unique internal links.")
-            return links
-        except Exception as e:
-            print(f"[!] Error extracting links: {e}")
-            return set()
-
-    def start_scan(self):
-        """Run vulnerability scans."""
-        print(f"[*] Starting scan for {self.target_url}...")
-        with sync_playwright() as p:
-            browser = p.chromium.launch(headless=True)
-            context = browser.new_context()
-            page = context.new_page()
-
-            try:
-                page.goto(self.target_url)
-                
-                if self.tests.get("check_http_headers"):
-                    self.check_http_headers(page)
-                
-                login_form = None
-                if self.tests.get("test_default_credentials"):
-                    login_form = self.find_login_page(page)
-                    if login_form:
-                        self.test_default_credentials(page, login_form)
-
-                links = []
-                if self.tests.get("extract_links"):
-                    links = self.extract_links(page)
-
-                if self.tests.get("find_input_elements"):
-                    for link in links:
-                        print(f"\n[*] Scanning linked route: {link}")
-                        page.goto(link)
-                        self.find_input_elements(page)
-
-            except Exception as e:
-                print(f"[!] Error during scan: {e}")
-            finally:
-                browser.close()
-
-def load_config(config_path):
-    try:
-        with open(config_path, 'r') as f:
-            return json.load(f)
-    except Exception as e:
-        print(f"[!] Error loading configuration file: {e}")
-        return {}
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(description="Web Vulnerability Scanner")
-    parser.add_argument("url", help="Target URL to scan (e.g., http://example.com)")
-    parser.add_argument(
-    "-c", "--config", help="Path to the configuration file (JSON format)", required=False)
-    args = parser.parse_args()
-
-    config = {}
-    if args.config:
-        config = load_config(args.config)
-    
-    scanner = WebVulnerabilityScanner(args.url, config=config)
-    scanner.start_scan()
